+++ myokit/_model_api.py
@@ -720,9 +720,7 @@ class VarOwner(ModelPart, VarProvider):
         """ See :meth:`VarProvider._resolve(). """
         def sa(name):
             # Suggest alternative
-            m = self.model()
-            (var, sug, msg) = m.suggest_variable(name)
-            return msg
+            return self.model().suggest_variable(name)[2]
 
         # Try resolving as an alias
         try:
@@ -772,7 +770,7 @@ class Model(ObjectWithMeta, VarProvider):
     Variables stored inside components can be accessed using :meth:`get()` or
     :meth:`values()`. Values defined through their derivative make up the
     model state and can be accessed using :meth:`states()`. States have
-    initial values accessible through :meth:`inits()`.
+    initial values accessible through :meth:`initial_values()`.
 
     A model's validity can be checked using :meth:`is_valid()`, which returns
     the latest validation status and :meth:`validate()`, which (re)validates
@@ -797,10 +795,10 @@ class Model(ObjectWithMeta, VarProvider):
         self._components = {}
 
         # The model's state variables
-        self._state = []
+        self._state_vars = []
 
-        # The model's current state (list of myokit.Expression)
-        self._current_state = []
+        # The model's initial state, as a list of Expressions
+        self._state_init = []
 
         # A dict mapping binding names to variables
         self._bindings = {}
@@ -1088,7 +1086,7 @@ class Model(ObjectWithMeta, VarProvider):
             c._clone1(clone)
 
         # Clone state
-        for k, v in enumerate(self._state):
+        for k, v in enumerate(self._state_vars):
             clone.get(v.qname()).promote()
 
         # Create mapping of old var references to new references
@@ -1111,10 +1109,9 @@ class Model(ObjectWithMeta, VarProvider):
             clone.reserve_unique_name_prefix(prefix, prepend)
 
         # Copy initial state expressions
-        for k, v in enumerate(self._state):
-            clone.get(v.qname()).set_state_value(
-                self._current_state[k].clone(subst=lhs_map)
-            )
+        for k, v in enumerate(self._state_vars):
+            clone.get(v.qname()).set_initial_value(
+                self._state_init[k].clone(subst=lhs_map))
 
         # Return
         return clone
@@ -1147,17 +1144,15 @@ class Model(ObjectWithMeta, VarProvider):
         self._code_meta(b, 0)
 
         # Initial state
-        if self._state:
+        if self._state_vars:
             pre = t * TAB
             b.write(pre + '# Initial values\n')
-            names = [eq.lhs.code() for eq in self.inits()]
+            eqs = self.initial_values(as_equations=True)
+            names = [eq.lhs.code() for eq in eqs]
             n = max([len(name) for name in names])
-            names = iter(names)
-            for eq in self.inits():
-                name = next(names)
-                b.write(
-                    pre + name + ' ' * (n - len(name)) + ' = ' + eq.rhs.code()
-                    + '\n')
+            for eq, name in zip(eqs, names):
+                b.write(pre + name + ' ' * (n - len(name)) + ' = '
+                        + eq.rhs.code() + '\n')
             b.write(pre + '\n')
         else:
             # No initial state? Then add newline
@@ -1249,7 +1244,7 @@ class Model(ObjectWithMeta, VarProvider):
         """
         Returns the number of state variables in this model.
         """
-        return len(self._state)
+        return len(self._state_vars)
 
     def create_unique_names(self):
         """
@@ -1363,7 +1358,7 @@ class Model(ObjectWithMeta, VarProvider):
         # Insert new state (if required)
         if state is not None:
             new_state = self.map_to_state(state)
-            for state, value in zip(self._state, new_state):
+            for state, value in zip(self._state_vars, new_state):
                 values[myokit.Name(state)] = value
             state = None
 
@@ -1396,7 +1391,7 @@ class Model(ObjectWithMeta, VarProvider):
                     values[eq.lhs] = eq.rhs.eval(values, precision=precision)
 
         # Return calculated state
-        return [values[state.lhs()] for state in self._state]
+        return [values[state.lhs()] for state in self._state_vars]
 
     def eval_state_derivatives(
             self, state=None, inputs=None, precision=myokit.DOUBLE_PRECISION,
@@ -1495,8 +1490,8 @@ class Model(ObjectWithMeta, VarProvider):
     def format_state(self, state=None, state2=None,
                      precision=myokit.DOUBLE_PRECISION):
         """
-        Converts the given list of floating point numbers to a string where
-        each line has the format ``<full_qualified_name> = <float_value>``.
+        Converts a sequence of floating point numbers to a string where each
+        line has the format ``<full_qualified_name> = <float_value>``.
 
         Arguments:
 
@@ -1511,19 +1506,19 @@ class Model(ObjectWithMeta, VarProvider):
             :meth:`myokit.float.str` when formatting the state values.
 
         """
-        n = len(self._state)
-        if state is not None:
+        n = len(self._state_vars)
+        if state is None:
+            state = self.initial_values(as_floats=True)
+        else:
             if len(state) != n:
                 raise ValueError(
-                    'Argument `state` must be a list of (' + str(n)
+                    'Argument `state` must be a sequence of (' + str(n)
                     + ') floating point numbers.')
-        else:
-            state = self.state_values()
 
         if state2 is not None:
             if len(state2) != n:
                 raise ValueError(
-                    'Argument `state2` must be a list of (' + str(n)
+                    'Argument `state2` must be a sequence of (' + str(n)
                     + ') floating point numbers.')
 
         out = []
@@ -1547,35 +1542,34 @@ class Model(ObjectWithMeta, VarProvider):
         Like :meth:`format_state` but displays the derivatives along with
         each state's value.
 
-
         Arguments:
 
         ``state=None``
             The state to display. If no state is given the state returned by
-            :meth:`state` is used.
+            :meth:`initial_values` is used.
         ``derivatives=None``
-            An optional list of evaluated derivatives. If not given, the values
-            will be calculed from ``state`` using :meth:`eval_derivatives()`.
+            An optional list or other sequence of evaluated derivatives. If not
+            given, the values will be calculed from ``state`` using
+            :meth:`eval_derivatives()`.
         ``precision=myokit.DOUBLE_PRECISION``
             An optional precision argument to use when evaluating the state
             derivatives, and to pass into :meth:`myokit.float.str` when
             formatting the state values and derivatives.
 
         """
-        n = len(self._state)
+        n = len(self._state_vars)
         if state is None:
-            state = self.state_values()
+            state = self.initial_values()
         elif len(state) != n:
             raise ValueError(
-                'Argument `state` must be a list of (' + str(n)
+                'Argument `state` must be a sequence of (' + str(n)
                 + ') floating point numbers.')
 
         if derivatives is None:
-            derivatives = self.evaluate_derivatives(
-                state, precision=precision)
+            derivatives = self.evaluate_derivatives(state, precision=precision)
         elif len(derivatives) != n:
             raise ValueError(
-                'Argument `deriv` must be a list of (' + str(n)
+                'Argument `deriv` must be a sequence of (' + str(n)
                 + ') floating point numbers.')
 
         out = []
@@ -1977,7 +1971,7 @@ class Model(ObjectWithMeta, VarProvider):
                 # Clone states
                 # TODO: Not sure why clone() code doesn't do this?
                 new_component[i].get(var.qname(comp)).promote(
-                    var.state_value())
+                    var.initial_value())
             # Now we can add variable to var_map if needed
             for var in map_to_clone[i]:
                 var_map[var] = new_component[i].get(var.qname(comp))
@@ -2037,15 +2031,40 @@ class Model(ObjectWithMeta, VarProvider):
                         var.set_rhs(
                             myokit.Multiply(rhs, myokit.Number(time_factor)))
 
-    def inits(self):
+    def initial_values(self, as_floats=False, as_equations=False):
         """
-        Returns an iterator over the ``Equation`` objects defining this model's
-        current state.
+        Returns a list with the model's initial state.
+
+        By default, expressions are returned, but this can be changed by
+        setting ``as_floats=True`` to obtain a list of floats, or
+        ``as_equations`` to obtain a list of equations.
         """
-        def StateDefIterator(model):
-            for var, state in zip(model._state, self._current_state):
-                yield Equation(myokit.Name(var), state)
-        return StateDefIterator(self)
+        if as_floats and as_equations:
+            raise ValueError('The arguments as_floats and as_equations cannot'
+                             ' both be set to True.')
+
+        if as_equations:
+            return [myokit.Equation(myokit.Name(x), y) for x, y in zip(
+                self._state_vars, self._state_init)]
+        if as_floats:
+            # Check for cyclical dependencies: only needed when evaluating
+            self.validate()
+
+            return [float(y) for y in self._state_init]
+        return list(self._state_init)
+
+    def OLD_inits(self):
+        """
+        Deprecated method, use
+        :meth:`iter(model.initial_values(as_equations=True))<initial_values>`
+        instead.
+        """
+        # Deprecated since 2022-11-24
+        import warnings
+        warnings.warn(
+            'The method `inits` is deprecated. Please use'
+            ' `iter(model.initial_values(as_equations=True)`.')
+        return iter(model.initial_values(as_equations=True))
 
     def is_similar(self, other, check_unames=False):
         """
@@ -2139,10 +2158,10 @@ class Model(ObjectWithMeta, VarProvider):
 
     def load_state(self, filename):
         """
-        Sets the model state using data from a file formatted in any style
-        accepted by :func:`myokit.parse_state`.
+        Sets the model's initial values using data from a file formatted in any
+        style accepted by :func:`myokit.map_to_state`.
         """
-        self.set_state(myokit.load_state(filename, self))
+        self.set_initial_values(myokit.load_state(filename, self))
 
     def map_component_dependencies(
             self, omit_states=True, omit_constants=False):
@@ -2706,7 +2725,7 @@ class Model(ObjectWithMeta, VarProvider):
 
         This method does not affect the model's validation status.
         """
-        n = len(self._state)
+        n = len(self._state_vars)
         if len(order) != n:
             raise ValueError(
                 'The given list must contain the same number of entries as'
@@ -2724,9 +2743,9 @@ class Model(ObjectWithMeta, VarProvider):
                     'Duplicate entry in order specification: "'
                     + str(v.qname()) + '".')
             state.append(v)
-            current.append(self._current_state[v._indice])
-        self._state = state
-        self._current_state = current
+            current.append(self._state_init[v._indice])
+        self._state_vars = state
+        self._state_init = current
         for k, v in enumerate(state):
             v._indice = k
 
@@ -2776,7 +2795,7 @@ class Model(ObjectWithMeta, VarProvider):
             time_unit = time.unit()
 
         # Scan all states
-        for state in self._state:
+        for state in self._state_vars:
 
             # Search for references to dot(state)
             refs = list(state.refs_by())
@@ -2847,7 +2866,7 @@ class Model(ObjectWithMeta, VarProvider):
         """
         Resets the indices of this model's state variables.
         """
-        for k, v in enumerate(self._state):
+        for k, v in enumerate(self._state_vars):
             v._indice = k
 
     def _reset_validation(self):
@@ -2858,14 +2877,11 @@ class Model(ObjectWithMeta, VarProvider):
 
     def _resolve(self, name):
         """ See :meth:`VarProvider._resolve(). """
-        def sa(name):
-            # Suggest alternative
-            (var, sug, msg) = self.suggest_variable(name)
-            return msg
         try:
             return self.get(name)
         except KeyError:
-            raise myokit.UnresolvedReferenceError(name, sa(name))
+            raise myokit.UnresolvedReferenceError(
+                name, self.suggest_variable(name)[2])
 
     def resolve_interdependent_components(self):
         """
@@ -2899,9 +2915,41 @@ class Model(ObjectWithMeta, VarProvider):
 
     def save_state(self, filename):
         """
-        Saves the model state to a file.
+        Saves the model state to a file (as floats).
+
+        Note that any expressions (either literal or containing references to
+        constants) will be evaluated and stored as numbers.
         """
-        return myokit.save_state(filename, self.state_values(), self)
+        return myokit.save_state(
+            filename, self.initial_values(as_floats=True), self)
+
+    def set_initial_values(self, state):
+        """
+        Sets this model's initial values using either a list of expressions or
+        anything that's accepted by :meth:`map_to_state`.
+        """
+        # Sequence of expressions?
+        is_expr = False
+        try:
+            if len(state) == len(self._state_vars):
+                if all(isinstance(i, myokit.Expression) for i in state):
+                    is_expr = True
+        except TypeError:
+            pass
+
+        # List of expressions
+        if is_expr:
+            for expr in state:
+                if not expr.is_constant():
+                    raise ValueError(
+                        'Expressions in a model\'s initial values may not'
+                        ' refer to variables that vary in time.')
+            self._state_init = list(state)
+
+        # Anything accepted by map_to_state
+        else:
+            self._state_init = [
+                myokit.Number(x) for x in self.map_to_state(state)]
 
     def set_name(self, name=None):
         """
@@ -2924,14 +2972,16 @@ class Model(ObjectWithMeta, VarProvider):
         self._reserved_unames = state[0]
         self._reserved_uname_prefixes = state[1]
 
-    def set_state(self, state):
+    def OLD_set_state(self, state):
         """
-        Changes this model's state. Accepts any type of input handled by
-        :meth:`map_to_state`.
+        Deprecated method: use :meth:`set_initial_values` instead.
         """
-        self._current_state = [
-            myokit.Number(x) for x in self.map_to_state(state)
-        ]
+        # Deprecated since 2022-11-24
+        import warnings
+        warnings.warn(
+            'The method `set_state` is deprecated. Please use'
+            ' `set_initial_values` instead.')
+        self.set_initial_values(state)
 
     def set_value(self, qname, value):
         """
@@ -2952,6 +3002,9 @@ class Model(ObjectWithMeta, VarProvider):
         The variable's equation and value are displayed, along with the value
         and formula of any nested variables and the values of all dependencies.
         """
+        # Model must be valid, or cycles can occur
+        self.model().validate()
+
         def format_float(number):
             s = str(number)
             if len(s) < 10:
@@ -2965,7 +3018,10 @@ class Model(ObjectWithMeta, VarProvider):
         # Add initial value
         rhs = var.rhs()
         if var.is_state():
-            out.append('Initial value = ' + str(var.state_value()))
+            value = var.initial_value()
+            out.append('Initial value = ' + value.code())
+            if not isinstance(value, myokit.Number):
+                out.append('              = ' + format_float(value))
             out.append(spacer)
         varname = var.lhs().code()
 
@@ -3249,27 +3305,24 @@ class Model(ObjectWithMeta, VarProvider):
         # Return
         return out
 
-    def state(self):
+    def OLD_state(self):
         """
-        Returns the current state of the model as a list of Expressions.
+        Deprecated method, use
+        :meth:`initial_values(as_floats=True)<initial_values>` instead.
         """
-        return [eqn.rhs for eqn in self.inits()]
-
-    def state_values(self):
-        """
-        Returns the current state of the model as a list of floating point
-        numbers.
-        """
-        # need to check for cyclic dependencies
-        self.validate()
-        return [float(eqn.rhs) for eqn in self.inits()]
+        # Deprecated since 2022-11-24
+        import warnings
+        warnings.warn(
+            'The method `state` is deprecated. Please use'
+            ' `initial_values(as_floats=True)` instead.')
+        return model.initial_values(as_floats=True)
 
     def states(self):
         """
         Returns an iterator over this model's state :class:`variable
         <myokit.Variable>` objects.
         """
-        return iter(self._state)
+        return iter(self._state_vars)
 
     def suggest_variable(self, name):
         """
@@ -3350,10 +3403,7 @@ class Model(ObjectWithMeta, VarProvider):
         "time". For a valid model, this method always returns a unique
         variable. If no time variable has been declared ``None`` is returned.
         """
-        try:
-            return self._bindings['time']
-        except KeyError:
-            return None
+        return self._bindings.get('time')
 
     def timex(self):
         """
@@ -3442,8 +3492,8 @@ class Model(ObjectWithMeta, VarProvider):
                     + n + '>.')
 
         # Test current state values
-        n = len(self._state)
-        if n != len(self._current_state):   # pragma: no cover
+        n = len(self._state_vars)
+        if n != len(self._state_init):   # pragma: no cover
             # Cover pragma: This can only happen if there's an API bug
             self._valid = False
             raise myokit.IntegrityError(
@@ -3457,12 +3507,11 @@ class Model(ObjectWithMeta, VarProvider):
         self.create_unique_names()
 
         # Check initial state expressions are still constant
-        for eqn in self.inits():
-            if not eqn.rhs.is_constant():
+        for var, expr in zip(self._state_vars, self._state_init):
+            if not expr.is_constant():
                 raise myokit.NonConstantExpressionError(
-                    'Initial condition for variable {} is '
-                    'not constant ({})'.format(eqn.lhs, eqn)
-                )
+                    'Initial value for variable <' + var.qname() + '> is not'
+                    ' constant: ' + expr.code() + '.')
 
         # Return
         self._valid = True
@@ -3511,7 +3560,7 @@ class Model(ObjectWithMeta, VarProvider):
 
         # Follow all state variables (unless already visited), all bound
         # variables and all used variables.
-        used = [x for x in self._state]
+        used = [x for x in self._state_vars]
         used += [x for x in self._bindings.values()]
         used += [x for x in self._labels.values()]
 
@@ -3913,9 +3962,15 @@ class Variable(VarOwner):
         - Any child variables will be removed.
 
         """
+        # Model must be valid before evaluations
+        self.model().validate()
+
         # Set value
         if value is None:
-            value = self.state_value() if self._is_state else self._rhs.eval()
+            if self._is_state:
+                value = self.initial_value().eval()
+            else:
+                value = self._rhs.eval()
         else:
             value = float(value)
 
@@ -4111,7 +4166,12 @@ class Variable(VarOwner):
 
         # For states, update the current/initial value
         if self._is_state:
-            self.set_state_value(self.state_value() * float(fw))
+            # Number? Then just multiply
+            value = self.initial_value()
+            if isinstance(value, myokit.Number):
+                self.set_initial_value(float(value) * float(fw))
+            else:
+                self.set_initial_value(myokit.Multiply(value, fw))
 
         # Update all references to the variable
         old_ref = myokit.Name(self)
@@ -4248,10 +4308,10 @@ class Variable(VarOwner):
         model = self.model()
         try:
             # Remove initial value
-            del model._current_state[self._indice]
+            del model._state_init[self._indice]
 
             # Remove this variable from the state
-            del model._state[self._indice]
+            del model._state_vars[self._indice]
 
             # Set lhs to name expression
             self._lhs = myokit.Name(self)
@@ -4298,6 +4358,21 @@ class Variable(VarOwner):
         if self._indice is None:
             raise Exception('Only state variables have initial values.')
         return self._indice
+
+    def initial_value(self, as_float=False):
+        """
+        Returns a state variable's initial value, or raises an exception when
+        called on a non-state variable.
+
+        By default, a :class:`myokit.Expression` is returned. To evaluate as a
+        float set ``as_float=True``.
+        """
+        if not self._is_state:
+            raise Exception('Only state variables have initial values.')
+        expr = self.model()._state_init[self._indice]
+        print(self, expr, type(expr))
+
+        return float(expr) if as_float else expr
 
     def is_bound(self):
         """
@@ -4371,24 +4446,25 @@ class Variable(VarOwner):
         """
         return self._lhs
 
-    def promote(self, state_value=0):
+    def promote(self, initial_value=0):
         """
-        Turns this variable into a state variable with a current state value
-        given by ``state_value``.
+        Turns this variable into a state variable with an initial value given
+        by ``initial_value``.
 
-        The new ``state_value`` should be:
+        The new ``initial_value`` should be:
 
-        1. a numerical value
-        2. a :class:`myokit.Expression`. If an expression is used, it can
-           contain references to model variables, as long as they are
-           constant-valued.
-        3. a string, in which case it is parsed to a
-           :class:`myokit.Expression`. Note that this parsing is done in the
-           context of the model, not the variable, so child variables cannot
-           be used
+        1. A numerical value.
+        2. A :class:`myokit.Expression`.
+        3. A string which can be parsed to a :class:`myokit.Expression`. Any
+           references to variables must be made using their fully qualified
+           names.
 
-        This will reset the validation status of the model this variable
-        belongs to.
+        Note that expressions can contain references to non-nested and
+        constant-valued variables (i.e. their right-hand side is either a
+        literal expression or refers only to constants).
+
+        Calling ``promote`` will reset the validation status of the model this
+        variable belongs to.
         """
         if self._indice is not None:
             raise Exception('Variable is already a state variable')
@@ -4399,36 +4475,33 @@ class Variable(VarOwner):
                 'State variables cannot be bound to an external value.')
 
         # Handle string and number rhs's
-        # expressions are evaluated in model context, not variables
-        if not isinstance(state_value, myokit.Expression):
-            if isinstance(state_value, basestring):
-                state_value = myokit.parse_expression(
-                    state_value, context=self.model()
-                )
-            elif state_value is not None:
-                state_value = myokit.Number(state_value)
-
         model = self.model()
+        if not isinstance(initial_value, myokit.Expression):
+            if isinstance(initial_value, basestring):
+                # Expressions are evaluated in model context
+                initial_value = myokit.parse_expression(
+                    initial_value, context=model)
+            elif initial_value is not None:
+                initial_value = myokit.Number(initial_value)
 
         # check initial state value expression is constant
-        if not state_value.is_constant():
+        if not initial_value.is_constant():
             raise myokit.NonConstantExpressionError(
-                'Expressions for state values must only contain '
-                'references to constant variables'
-            )
+                'Expressions for state values must only contain references to'
+                ' constant variables.')
 
         try:
             # Set lhs to derivative expression
             self._lhs = myokit.Derivative(myokit.Name(self))
 
             # Get new indice
-            self._indice = len(model._state)
+            self._indice = len(model._state_vars)
 
             # Add to list of states
-            model._state.append(self)
+            model._state_vars.append(self)
 
-            # Add state_value to list of current values
-            model._current_state.append(state_value)
+            # Add initial_value to list of current values
+            model._state_init.append(initial_value)
 
             # All references to this variable are now considered references to
             # its state value
@@ -4658,6 +4731,36 @@ class Variable(VarOwner):
             # Reset model validation
             model._reset_validation()
 
+    def set_initial_value(self, value):
+        """
+        Sets the initial value of a state variable, or raises an exception if
+        called on a non-state variable.
+
+        The new value can be passed in as an expression, number, or a string
+        (in which case it will be parsed as an expression). Expressions can
+        refer to variables as long as they are not nested and are constant in
+        time. Variable references in strings must be made using fully qualified
+        names (``component.variable``).
+        """
+        if not self._is_state:
+            raise Exception('Only state variables have state values.')
+
+        # Handle strings and floats
+        model = self.model()
+        if not isinstance(value, myokit.Expression):
+            if isinstance(value, basestring):
+                value = myokit.parse_expression(value, context=model)
+            else:
+                value = myokit.Number(value)
+
+        # Check expression
+        if not value.is_constant():
+            raise myokit.NonConstantExpressionError(
+                'Expressions for state values must be constant in time.')
+
+        model._state_init[self._indice] = value
+        # No need to reset validation status or cache here.
+
     def set_label(self, label=None):
         """
         Adds a unique ``label`` for this variable, indicated that its value can
@@ -4743,23 +4846,16 @@ class Variable(VarOwner):
         self.model()._reset_validation()
         self._reset_cache(bubble=True)
 
-    def set_state_value(self, value):
+    def OLD_set_state_value(self, value):
         """
-        If this variable is a state variable, its current value will be
-        updated. For all other variables this raises an exception.
+        Deprecated method, use :meth:`set_initial_value` instead.
         """
-        if not self._is_state:
-            raise Exception('Only state variables have state values.')
-        model = self.model()
-        if isinstance(value, myokit.Expression):
-            if not value.is_constant():
-                raise myokit.NonConstantExpressionError(
-                    'Expressions for state values must be constant in time.')
-            expr = value
-        else:
-            expr = myokit.Number(float(value))
-        model._current_state[self._indice] = expr
-        # No need to reset validation status or cache here.
+        # Deprecated since 2022-11-24
+        import warnings
+        warnings.warn(
+            'The method `set_state_value` is deprecated. Please use'
+            ' `set_initial_value` instead.')
+        self.set_initial_value(value)
 
     def set_unit(self, unit=None):
         """
@@ -4775,14 +4871,17 @@ class Variable(VarOwner):
         # No need to reset validation status or cache. Units are checked only
         # by the model.
 
-    def state_value(self):
+    def OLD_state_value(self):
         """
-        For state variables, this will return their current value.
-        For all other variables, this will raise an exception.
+        Deprecated method, use
+        :meth:`initial_value(as_float=True)<initial_value>` instead.
         """
-        if not self._is_state:
-            raise Exception('Only state variables have initial values.')
-        return float(self.model()._current_state[self._indice])
+        # Deprecated since 2022-11-24
+        import warnings
+        warnings.warn(
+            'The method `state_value` is deprecated. Please use'
+            ' `initial_value(as_float=True)` instead.')
+        return self.initial_value(as_float=True)
 
     def unit(self, mode=myokit.UNIT_TOLERANT):
         """
@@ -4842,7 +4941,7 @@ class Variable(VarOwner):
                     'State variables should not be nested: <'
                     + str(self.qname()) + '>.')
             m = self.model()
-            if not m._state[self._indice] == self:  # pragma: no cover
+            if not m._state_vars[self._indice] == self:  # pragma: no cover
                 raise myokit.IntegrityError(
                     'State variable not listed in model state vector at'
                     ' correct indice: <' + self.qname() + '>.')